requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
numpy==1.24.3
tensorflow==2.17.0
Pillow==10.0.0